ref: Move SymbolSupplier out of the processor
The SymbolSupplier was so far re-exported from breakpad-symbols along
with some constructor functions.
